+++ src/main/java/com/timemanagemenet/timemanagementapp/Controller/AttendanceController.java
@@ -8,7 +8,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -26,6 +28,14 @@ public class AttendanceController {
     @GetMapping
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     public ResponseEntity<List<Attendance>> getAllAttendances() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        boolean isAdmin = authentication.getAuthorities().stream()
+                .anyMatch(authority -> authority.getAuthority().equals("ROLE_ADMIN"));
+
+        if (!isAdmin) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         List<Attendance> attendances = attendanceService.getAllAttendances();
         return ResponseEntity.ok(attendances);
     }
@@ -33,6 +43,14 @@ public class AttendanceController {
     @GetMapping("/{attendanceId}")
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     public ResponseEntity<Attendance> getAttendanceById(@PathVariable Long attendanceId) throws NotFoundException {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        boolean isAdmin = authentication.getAuthorities().stream()
+                .anyMatch(authority -> authority.getAuthority().equals("ROLE_ADMIN"));
+
+        if (!isAdmin) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         Attendance attendance = attendanceService.getAttendanceById(attendanceId);
         return ResponseEntity.ok(attendance);
     }
@@ -52,6 +70,14 @@ public class AttendanceController {
     @PostMapping
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     public ResponseEntity<Attendance> createAttendance(@RequestBody Attendance attendance) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        boolean isAdmin = authentication.getAuthorities().stream()
+                .anyMatch(authority -> authority.getAuthority().equals("ROLE_ADMIN"));
+
+        if (!isAdmin) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         Attendance createdAttendance = attendanceService.createAttendance(attendance);
         return ResponseEntity.status(HttpStatus.CREATED).body(createdAttendance);
     }
@@ -61,6 +87,14 @@ public class AttendanceController {
     public ResponseEntity<Attendance> updateAttendance(
             @PathVariable Long attendanceId,
             @RequestBody Attendance updatedAttendance) throws NotFoundException {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        boolean isAdmin = authentication.getAuthorities().stream()
+                .anyMatch(authority -> authority.getAuthority().equals("ROLE_ADMIN"));
+
+        if (!isAdmin) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         Attendance updated = attendanceService.updateAttendance(attendanceId, updatedAttendance);
         return ResponseEntity.ok(updated);
     }
@@ -68,6 +102,14 @@ public class AttendanceController {
     @DeleteMapping("/{attendanceId}")
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     public ResponseEntity<Void> deleteAttendance(@PathVariable Long attendanceId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        boolean isAdmin = authentication.getAuthorities().stream()
+                .anyMatch(authority -> authority.getAuthority().equals("ROLE_ADMIN"));
+
+        if (!isAdmin) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         attendanceService.deleteAttendance(attendanceId);
         return ResponseEntity.noContent().build();
     }
